+++ src/bundler/WebpackBundler.js
@@ -13,7 +13,7 @@ import { fileURLToPath } from 'url';
 import path from 'path';
 import fse from 'fs-extra';
 import webpack from 'webpack';
-import chalk from 'chalk';
+import chalk from 'chalk-template';
 import BaseBundler from './BaseBundler.js';
 
 // eslint-disable-next-line no-underscore-dangle
@@ -74,6 +74,8 @@ export default class WebpackBundler extends BaseBundler {
           // the main.js is imported in the universal adapter and is _the_ action entry point
           './main.js': cfg.file,
         },
+        // use fixed conditions to omit the `development` condition.
+        conditionNames: ['node', 'require'],
       },
       node: {
         __dirname: true,