+++ src/deploy/OpenWhiskDeployer.js
@@ -14,7 +14,7 @@ import ow from 'openwhisk';
 import os from 'os';
 import fse from 'fs-extra';
 import path from 'path';
-import chalk from 'chalk';
+import chalk from 'chalk-template';
 import dotenv from 'dotenv';
 import BaseDeployer from './BaseDeployer.js';
 import OpenWhiskConfig from './OpenWhiskConfig.js';
@@ -86,7 +86,7 @@ export default class OpenWhiskDeployer extends BaseDeployer {
 
   getOpenwhiskClient() {
     if (!this._cfg.apiHost || !this._cfg.auth || !this._cfg.namespace) {
-      throw Error(chalk`\nMissing OpenWhisk credentials. Make sure you have a {grey .wskprops} in your home directory.\nYou can also set {grey WSK_NAMESPACE}, {gray WSK_AUTH} and {gray WSK_API_HOST} environment variables.`);
+      throw Error(chalk`\nMissing OpenWhisk credentials. Make sure you have a {grey .wskprops} in your home directory.\nYou can also set {grey WSK_NAMESPACE}, {grey WSK_AUTH} and {grey WSK_API_HOST} environment variables.`);
     }
     return ow({
       apihost: this._cfg.apiHost,
@@ -115,7 +115,7 @@ export default class OpenWhiskDeployer extends BaseDeployer {
         });
         this.log.info(chalk`{green ok:} package created. ${res.namespace}/${res.name}`);
       } else {
-        this.log.error(`${chalk.red('error: ')} ${e.message}`);
+        this.log.error(chalk`{red error:} ${e.message}`);
       }
     }
 
@@ -185,7 +185,7 @@ export default class OpenWhiskDeployer extends BaseDeployer {
         fn = openwhisk.packages.create.bind(openwhisk.packages);
         verb = 'created';
       } else {
-        this.log.error(`${chalk.red('error: ')} ${e.message}`);
+        this.log.error(chalk`{red error:} ${e.message}`);
       }
     }
 
@@ -201,9 +201,9 @@ export default class OpenWhiskDeployer extends BaseDeployer {
           parameters,
         },
       });
-      this.log.info(`${chalk.green('ok:')} ${verb} package ${chalk.whiteBright(`/${result.namespace}/${result.name}`)}`);
+      this.log.info(chalk`{green ok:} ${verb} package {whiteBright /${result.namespace}/${result.name}}`);
     } catch (e) {
-      this.log.error(`${chalk.red('error: failed processing package: ')} ${e.stack}`);
+      this.log.error(chalk`{red error: failed processing package: } ${e.stack}`);
       throw Error('abort.');
     }
   }
@@ -268,10 +268,10 @@ export default class OpenWhiskDeployer extends BaseDeployer {
       try {
         this.log.debug(`creating sequence: ${options.name} -> ${options.action.exec.components[0]}`);
         const result = await openwhisk.actions.update(options);
-        this.log.info(`${chalk.green('ok:')} created sequence ${chalk.whiteBright(`/${result.namespace}/${result.name}`)} -> ${chalk.whiteBright(fqn)}`);
+        this.log.info(chalk`{green ok:} created sequence {whiteBright /${result.namespace}/${result.name}} -> {whiteBright ${fqn}}`);
       } catch (e) {
         hasErrors = true;
-        this.log.error(`${chalk.red('error:')} failed creating sequence: ${e.message}`);
+        this.log.error(chalk`{red error:} failed creating sequence: ${e.message}`);
       }
     }));
     if (hasErrors) {