+++ test/fixtures/simple/index.js
@@ -14,6 +14,9 @@ const reader = require('./helper/read.js');
 
 // eslint-disable-next-line no-unused-vars
 module.exports.main = function main(req, context) {
+  if (req.url.endsWith('/fail')) {
+    throw new Error('internal error');
+  }
   const url = new URL(req.url);
   const chanceoffailure = parseInt(url.searchParams.get('chanceoffailure') || '0', 10);
   if (Math.random() * 100 < chanceoffailure) {